fix: purge storage on schema validatione errors + cleanup
