+++ packages/wallet-lib/src/types/ChainStore/ChainStore.js
@@ -4,23 +4,6 @@ const {
   Transaction, BlockHeader,
 } = require('@dashevo/dashcore-lib');
 
-const SCHEMA = {
-  '*': {
-    blockHeaders: {
-      '*': function (hex) { return new BlockHeader(Buffer.from(hex, 'hex')); },
-    },
-    transactions: {
-      '*': Transaction,
-    },
-    txMetadata: {
-      '*': {
-        blockHash: 'string',
-        height: 'number',
-      },
-    },
-  },
-};
-
 /**
  * ChainStore holds any information that is relatives to a specific network.
  * Information such as blockHeaders, transactions, instantLocks.
@@ -47,8 +30,6 @@ class ChainStore extends EventEmitter {
     return this.state.transactions;
   }
 }
-
-ChainStore.SCHEMA = SCHEMA;
 
 ChainStore.prototype.considerTransaction = require('./methods/considerTransaction');
