+++ packages/wallet-lib/src/types/Storage/methods/rehydrateState.js
@@ -1,40 +1,11 @@
-const { hasMethod } = require('../../../utils');
+const {hasMethod} = require('../../../utils');
 
-const { REHYDRATE_STATE_FAILED, REHYDRATE_STATE_SUCCESS } = require('../../../EVENTS');
+const {REHYDRATE_STATE_FAILED, REHYDRATE_STATE_SUCCESS} = require('../../../EVENTS');
 
 const logger = require('../../../logger');
 const WalletStore = require('../../WalletStore/WalletStore');
 const ChainStore = require('../../ChainStore/ChainStore');
 
-const castItemTypes = (item, schema) => {
-  Object.entries(schema).forEach(([schemaKey, schemaValue]) => {
-    if (schemaValue.constructor.name !== 'Object') {
-      const Clazz = schemaValue;
-      if (schemaKey === '*') {
-        Object.keys(item).forEach((itemKey) => {
-          // eslint-disable-next-line no-param-reassign
-          item[itemKey] = new Clazz(item[itemKey]);
-        });
-      } else {
-        if (!item[schemaKey]) {
-          throw new Error(`No schema key "${schemaKey}" found for item ${JSON.stringify(item)}`);
-        }
-
-        // todo typeof
-        if (!(['string', 'number'].includes(schemaValue))) {
-          // eslint-disable-next-line no-param-reassign
-          item[schemaKey] = new Clazz(item[schemaKey]);
-        }
-      }
-    } else if (schemaKey === '*') {
-      Object.values(item).forEach((itemValue) => castItemTypes(itemValue, schemaValue));
-    } else {
-      castItemTypes(item[schemaKey], schemaValue);
-    }
-  });
-
-  return item;
-};
 
 /**
  * Fetch the state from the persistence adapter
@@ -46,12 +17,6 @@ const rehydrateState = async function rehydrateState() {
       if (this.adapter && hasMethod(this.adapter, 'getItem')) {
         const wallets = await this.adapter.getItem('wallets');
         if (wallets) {
-          try {
-            castItemTypes(wallets, WalletStore.SCHEMA);
-          } catch (e) {
-            this.adapter.setItem('wallets', null);
-          }
-
           Object.keys(wallets).forEach((walletId) => {
             const walletStore = this.getWalletStore(walletId);
             if (walletStore) {
@@ -62,12 +27,6 @@ const rehydrateState = async function rehydrateState() {
 
         const chains = await this.adapter.getItem('chains');
         if (chains) {
-          try {
-            castItemTypes(chains, ChainStore.SCHEMA);
-          } catch (e) {
-            this.adapter.setItem('chains', null);
-          }
-
           Object.keys(chains).forEach((chainNetwork) => {
             const chainStore = this.getChainStore(chainNetwork);
             if (chainStore) {
