+++ packages/wallet-lib/src/types/WalletStore/WalletStore.js
@@ -1,11 +1,3 @@
-const SCHEMA = {
-  '*': {
-    lastKnownBlock: {
-      height: 'number',
-    },
-  },
-};
-
 class WalletStore {
   constructor(walletId) {
     this.walletId = walletId;
@@ -32,8 +24,6 @@ class WalletStore {
     this.state.lastKnownBlock.height = height;
   }
 }
-
-WalletStore.SCHEMA = SCHEMA;
 
 WalletStore.prototype.createPathState = require('./methods/createPathState');
 WalletStore.prototype.exportState = require('./methods/exportState');
